Slow down bubble oscillation; make bubbles scale grow from spawn point; templatise effect removal
