+++ TombEngine/Objects/TR3/Vehicles/rubber_boat.cpp
@@ -5,6 +5,7 @@
 #include "Game/camera.h"
 #include "Game/collision/collide_item.h"
 #include "Game/collision/sphere.h"
+#include "Game/effects/effects.h"
 #include "Game/effects/Bubble.h"
 #include "Game/items.h"
 #include "Game/Lara/lara.h"
